Split up the test API and SMTP server into two microservices
