+++ skyportal/services/test_server/supervisor.conf
@@ -1,5 +1,11 @@
-[program:testserver]
-command=/usr/bin/env python skyportal/services/test_server/test_server.py %(ENV_FLAGS)s
+[program:testapiserver]
+command=/usr/bin/env python skyportal/services/test_server/test_api_server.py %(ENV_FLAGS)s
 environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
-stdout_logfile=log/test_server.log
+stdout_logfile=log/test_api_server.log
+redirect_stderr=true
+
+[program:testsmtpserver]
+command=/usr/bin/env python skyportal/services/test_server/test_smtp_server.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+stdout_logfile=log/test_smtp_server.log
 redirect_stderr=true
